+++ src/Client.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Maxemail\Api;
 
 use GuzzleHttp\Client as GuzzleClient;
+use GuzzleHttp\ClientInterface as GuzzleClientInterface;
 use GuzzleHttp\HandlerStack;
 use Psr\Log\LoggerInterface;
 
@@ -68,6 +69,11 @@ class Client implements \Psr\Log\LoggerAwareInterface
     /**
      * @var string
      */
+    private $token;
+
+    /**
+     * @var string
+     */
     private $username;
 
     /**
@@ -91,7 +97,7 @@ class Client implements \Psr\Log\LoggerAwareInterface
     private $logger;
 
     /**
-     * @var GuzzleClient
+     * @var GuzzleClientInterface
      */
     private $httpClient;
 
@@ -99,6 +105,11 @@ class Client implements \Psr\Log\LoggerAwareInterface
      * @var bool
      */
     private $debugLoggingEnabled = false;
+
+    /**
+     * @var \Closure(array):GuzzleClientInterface
+     */
+    private $httpClientFactory;
 
     /**
      * @param array $config {
@@ -112,20 +123,25 @@ class Client implements \Psr\Log\LoggerAwareInterface
      */
     public function __construct(array $config)
     {
-        // Support deprecated key names from v3
-        if (!isset($config['username']) && isset($config['user'])) {
-            $config['username'] = $config['user'];
-        }
-        if (!isset($config['password']) && isset($config['pass'])) {
-            $config['password'] = $config['pass'];
-        }
+        // Must have API token
+        if (!isset($config['token'])) {
+            // Support deprecated key names from v3
+            if (!isset($config['username']) && isset($config['user'])) {
+                $config['username'] = $config['user'];
+            }
+            if (!isset($config['password']) && isset($config['pass'])) {
+                $config['password'] = $config['pass'];
+            }
 
-        // Must have user/pass
-        if (!isset($config['username']) || !isset($config['password'])) {
-            throw new Exception\InvalidArgumentException('API config requires username & password');
+            // Must have user/pass
+            if (!isset($config['username']) || !isset($config['password'])) {
+                throw new Exception\InvalidArgumentException('API config requires token OR username & password');
+            }
+            $this->username = $config['username'];
+            $this->password = $config['password'];
+        } else {
+            $this->token = $config['token'];
         }
-        $this->username = $config['username'];
-        $this->password = $config['password'];
 
         if (isset($config['uri'])) {
             $parsed = parse_url($config['uri']);
@@ -157,7 +173,7 @@ class Client implements \Psr\Log\LoggerAwareInterface
         return $this->services[$serviceName];
     }
 
-    private function getClient(): GuzzleClient
+    private function getClient(): GuzzleClientInterface
     {
         if ($this->httpClient === null) {
             $stack = HandlerStack::create();
@@ -166,19 +182,31 @@ class Client implements \Psr\Log\LoggerAwareInterface
             if ($this->debugLoggingEnabled) {
                 Middleware::addLogging($stack, $this->getLogger());
             }
-            $this->httpClient = new GuzzleClient([
+
+            $clientConfig = [
                 'base_uri' => $this->uri . 'api/json/',
-                'auth' => [
-                    $this->username,
-                    $this->password,
-                ],
                 'headers' => [
                     'User-Agent' => 'MxmApiClient/' . self::VERSION . ' PHP/' . PHP_VERSION,
                     'Content-Type' => 'application/x-www-form-urlencoded',
                     'Accept' => 'application/json',
                 ],
                 'handler' => $stack,
-            ]);
+            ];
+
+            if (isset($this->token)) {
+                $clientConfig['headers']['Authorization'] = 'Bearer ' . $this->token;
+            } else {
+                $clientConfig['auth'] = [
+                    $this->username,
+                    $this->password,
+                ];
+            }
+
+            if (!isset($this->httpClientFactory)) {
+                $this->httpClient = new GuzzleClient($clientConfig);
+            } else {
+                $this->httpClient = ($this->httpClientFactory)($clientConfig);
+            }
         }
 
         return $this->httpClient;
@@ -187,6 +215,7 @@ class Client implements \Psr\Log\LoggerAwareInterface
     /**
      * Get API connection config
      *
+     * @deprecated v5.2 No replacement; packages can maintain their own config; to be removed in v7.
      * @return array {
      *     @var string $uri
      *     @var string $username
@@ -223,5 +252,13 @@ class Client implements \Psr\Log\LoggerAwareInterface
         }
 
         return $this->logger;
+    }
+
+    /**
+     * @internal This method is not part of the BC promise. Used for DI for unit tests only.
+     */
+    public function setHttpClientFactory(\Closure $httpClientFactory): void
+    {
+        $this->httpClientFactory = $httpClientFactory;
     }
 }